Unify Town and Nation code, and introduce new Eco Account Mechanics (#3988)
* Unify Bank methods and provide defaults.

* Introduce a new generified event BankTransactionEvent for listening to *any* bank transaction.

* Make sure BankTransactionEvent has static handerlist.

* Fix misc javadoc errors that were leftover.

* Introduce new territory for code consolidation between towns and nations.

* Revert Towny.java changes

* More cleanup

* Add new SpawnLocation interface and more cleanup

* Cleanup Commit

Removes unused imports
Re-adds renamed methods as deprecated to maintain compatibility.

* Format code to what is standard in Towny.

* Add new Bank Object

* Add Documentation

* Cleanup

* Add more specific documentation.

* Add observer for account transactions to separate auditing code.

* Make most objects return unmodifiable lists, this will prevent untracked and dangerous writes to objects.

- Document ResidentList

* Remove some unused imports.

* Add auditing mechanics to Town and Nation

* Initial book stuff.

* Add iterator support for TownBlockOwner and ResidentList

* Add deprecations for backwards compatibility.

* Revert iterators.

* actually remove it

* Finalization.

* Fix version number

* Add economy adapter to interface with multiple eco implementations.

* unbox primitive types.

* fix the other primitives

* Remove GH packages dependencies.

* Attach sources

* Remove static server account field

* Fix NPE on startup

* Initial skeleton for debt handling.

* Flesh out more debt methods add debt prefix config setting.

* Debt account is now functional.

* make methods protected

* Fix deposit bug

* Fix other debt issues.

* remove unecessary stuff

* add docs

* Slight tweak to BankEconomyHandler

* one fix my dude

* fix javadoc errors.

* make requested changes.

* Add iterators because tb lists are gone.

* Add fixes from goosius' branch

* Cleanup/Territory renameRemove unused imports

* Fix tag issues

* Fix tne account name issue.

* Add initial proper eco bank support.

* Revert "Add initial proper eco bank support."

This reverts commit 9c838ae8

* Polish it up

* Refactor things that are useless and/or serve no purpose, and adjust naming.

* Fix this thang.

* Cool

* Cool

* oops ;(

* Update comparators for government types.

* Fix possible EconomyAccount deadlock issue.

* Remove commented code, add back missing exception.

* merge fixes

* fix merge error

* Merge upstream changes.

* Rename Residence to ResidentOwner

* Revert back to ResidentList.

* fix javadoc error.

* Removed unused imports.

* Clean up

- unused imports
- undo orphanged residentList->owner renames
- remove unused method in Town object.

+++ src/com/palmergames/bukkit/towny/object/Transaction.java
@@ -5,9 +5,9 @@ import org.bukkit.entity.Player;
 public class Transaction {
 	private final TransactionType type;
 	private final Player player;
-	private final int amount;
+	private final double amount;
 	
-	public Transaction(TransactionType type, Player player, int amount) {
+	public Transaction(TransactionType type, Player player, double amount) {
 		this.type = type;
 		this.player = player;
 		this.amount = amount;
@@ -21,7 +21,7 @@ public class Transaction {
 		return player;
 	}
 
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 }
